Created mappings for find queries

+++ group assign co/roadco/src/main/java/groupassign/roadco/model/repositories/ClientRepo.java
@@ -4,12 +4,14 @@ package groupassign.roadco.model.repositories;
 
 import java.util.List;
 
+
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import groupassign.roadco.model.entities.Client;
 
 @Repository
-public interface ClientRepo extends CrudRepository<Client,Long>{
+public interface ClientRepo extends CrudRepository<Client,Integer>{
 
     //Queries to find a client
     List<Client> findByClientCompany(String clientCompany);
@@ -17,6 +19,7 @@ public interface ClientRepo extends CrudRepository<Client,Long>{
     List<Client> findByClientCompanyAndClientAddress(String clientCompany, String clientAddress);
     //Queries to update a client
 
-    //Add a client
- 
+    
+
+    
 }
